pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-tenantkit"
version = "0.1.0"
description = "A comprehensive Django multi-tenancy solution with schema isolation, tenant-aware middleware, and flexible routing"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "Pablo Digonzelli", email = "pdigonzelli@gmail.com"}
]
keywords = ["django", "tenantkit", "multi-tenant", "saas", "postgresql", "schema"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "django>=6.0.3",
    "django-extensions>=4.1",
    "djangorestframework>=3.17.0",
    "drf-spectacular>=0.29.0",
    "psycopg[binary,pool]>=3.3.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "pyright>=1.1.0",
    "coverage>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/pdigonzelli/django-tenantkit"
Documentation = "https://github.com/pdigonzelli/django-tenantkit#readme"
Repository = "https://github.com/pdigonzelli/django-tenantkit"
Issues = "https://github.com/pdigonzelli/django-tenantkit/issues"
Changelog = "https://github.com/pdigonzelli/django-tenantkit/blob/main/CHANGELOG.md"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["tenantkit", "tenantkit.*"]

[tool.setuptools.package-data]
tenantkit = [
    "templates/**/*.html",
    "static/**/*.js",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py", "tests_*.py", "tests_integration.py", "tests_admin_base.py"]
testpaths = ["src/tenantkit"]
